requirements.txt
Flask==3.1.0
uuid==1.30
pillow==11.0.0
gunicorn==23.0.0
logging==0.4.9.6
redis==5.2.0
matplotlib>=3.0.0
numpy>=2.2.2